Cocoa: Working on CMake build system. Found compile issues that the native Xcode project was masking.

+++ src/cocoa/iupmac_globalattrib.m
@@ -26,7 +26,9 @@ static void macGlobalSendKey(int key, int press)
 {
 #if 0
   Ihandle* focus;
-  int keyval, state;
+ // original code used int for guint. anging to NSUInteger
+ //int keyval, state;
+ NSUInteger keyval, state;
 
   focus = IupGetFocus();
   if (!focus)
